RawBuffer: Convert PTS to 64-bit integer.
PTS parameter is used only in one place and that place treats pts as
an 64-bit value already so this change is totally safe.

+++ lmm/rawbuffer.h
@@ -30,7 +30,7 @@ public:
 	int h264NalType;
 	quintptr * omxBuf;
 	int poolIndex;
-	int pts;
+	qint64 pts;
 	int streamBufferNo;
 	quintptr *targetElement;
 	quintptr *v4l2Buffer;
